Revert "Merge pull request #52 from shijinglu/flatter_nest_submodules"
This reverts commit e28c075f0bda878a000ca40a3ab812ddf673a34e, reversing
changes made to db1bfefd41fed036660233e6574391f939af1d7d.
