+++ src/main/java/com/github/git24j/core/Blame.java
@@ -15,6 +15,102 @@ public class Blame extends CAutoReleasable {
         jniFree(cPtr);
     }
 
+    public enum FlagT implements IBitEnum {
+        /** Normal blame, the default */
+        NORMAL(0),
+        /** Track lines that have moved within a file (like `git blame -M`). NOT IMPLEMENTED. */
+        TRACK_COPIES_SAME_FILE(1 << 0),
+        /**
+         * Track lines that have moved across files in the same commit (like `git blame -C`). NOT
+         * IMPLEMENTED.
+         */
+        TRACK_COPIES_SAME_COMMIT_MOVES(1 << 1),
+        /**
+         * Track lines that have been copied from another file that exists in the same commit (like
+         * `git blame -CC`). Implies SAME_FILE. NOT IMPLEMENTED.
+         */
+        TRACK_COPIES_SAME_COMMIT_COPIES(1 << 2),
+        /**
+         * Track lines that have been copied from another file that exists in *any* commit (like
+         * `git blame -CCC`). Implies SAME_COMMIT_COPIES. NOT IMPLEMENTED.
+         */
+        TRACK_COPIES_ANY_COMMIT_COPIES(1 << 3),
+        /** Restrict the search of commits to those reachable following only the first parents. */
+        FIRST_PARENT(1 << 4),
+        /**
+         * Use mailmap file to map author and committer names and email addresses to canonical real
+         * names and email addresses. The mailmap will be read from the working directory, or HEAD
+         * in a bare repository.
+         */
+        USE_MAILMAP(1 << 5);
+        private final int _bit;
+
+        FlagT(int bit) {
+            _bit = bit;
+        }
+
+        @Override
+        public int getBit() {
+            return _bit;
+        }
+    }
+
+    public static class Options extends CAutoReleasable {
+        public static final int VERSION = 1;
+
+        protected Options(boolean isWeak, long rawPtr) {
+            super(isWeak, rawPtr);
+        }
+
+        @Override
+        protected void freeOnce(long cPtr) {
+            Libgit2.jniShadowFree(cPtr);
+        }
+
+        public static Options init(int version) {
+            Options out = new Options(false, 0);
+            jniOptionsNew(out._rawPtr, version);
+            return out;
+        }
+    }
+
+    /**
+     * Structure that represents a blame hunk.
+     *
+     * <pre>
+     * - `lines_in_hunk` is the number of lines in this hunk
+     * - `final_commit_id` is the OID of the commit where this line was last
+     *   changed.
+     * - `final_start_line_number` is the 1-based line number where this hunk
+     *   begins, in the final version of the file
+     * - `final_signature` is the author of `final_commit_id`. If
+     *   `GIT_BLAME_USE_MAILMAP` has been specified, it will contain the canonical
+     *    real name and email address.
+     * - `orig_commit_id` is the OID of the commit where this hunk was found.  This
+     *   will usually be the same as `final_commit_id`, except when
+     *   `GIT_BLAME_TRACK_COPIES_ANY_COMMIT_COPIES` has been specified.
+     * - `orig_path` is the path to the file where this hunk originated, as of the
+     *   commit specified by `orig_commit_id`.
+     * - `orig_start_line_number` is the 1-based line number where this hunk begins
+     *   in the file named by `orig_path` in the commit specified by
+     *   `orig_commit_id`.
+     * - `orig_signature` is the author of `orig_commit_id`. If
+     *   `GIT_BLAME_USE_MAILMAP` has been specified, it will contain the canonical
+     *    real name and email address.
+     * - `boundary` is 1 iff the hunk has been tracked to a boundary commit (the
+     *   root, or the commit specified in git_blame_options.oldest_commit)
+     * </pre>
+     */
+    public static class Hunk extends CAutoReleasable {
+        protected Hunk(boolean isWeak, long rawPtr) {
+            super(isWeak, rawPtr);
+        }
+
+        @Override
+        protected void freeOnce(long cPtr) {
+            Libgit2.jniShadowFree(cPtr);
+        }
+    }
 
     /** int git_blame_init_options(git_blame_options *opts, unsigned int version); */
     static native int jniInitOptions(long opts, int version);
@@ -39,12 +135,12 @@ public class Blame extends CAutoReleasable {
      * @return the hunk at the given index, or empty if non is available
      * @throws GitException git errors
      */
-    public Optional<BlameHunk> getHunkByIndex(int index) {
+    public Optional<Hunk> getHunkByIndex(int index) {
         long ptr = jniGetHunkByindex(getRawPointer(), index);
         if (ptr == 0) {
             return Optional.empty();
         }
-        return Optional.of(new BlameHunk(false, ptr));
+        return Optional.of(new Hunk(false, ptr));
     }
 
     /** const git_blame_hunk * git_blame_get_hunk_byline(git_blame *blame, size_t lineno); */
@@ -56,9 +152,9 @@ public class Blame extends CAutoReleasable {
      * @param lineno the (1-based) line number to find a hunk for
      * @return the hunk that contains the given line, or NULL on error
      */
-    public Optional<BlameHunk> getHunkByLine(int lineno) {
+    public Optional<Hunk> getHunkByLine(int lineno) {
         long ptr = jniGetHunkByline(getRawPointer(), lineno);
-        return ptr == 0 ? Optional.empty() : Optional.of(new BlameHunk(false, ptr));
+        return ptr == 0 ? Optional.empty() : Optional.of(new Hunk(false, ptr));
     }
 
     /**
@@ -78,7 +174,7 @@ public class Blame extends CAutoReleasable {
      */
     @Nonnull
     public static Blame file(
-            @Nonnull Repository repo, @Nonnull String path, @Nullable BlameOptions options) {
+            @Nonnull Repository repo, @Nonnull String path, @Nullable Options options) {
         Blame blame = new Blame(false, 0);
         Error.throwIfNeeded(
                 jniFile(
